Rename put ActivityCard in file with same name
