+++ src/components/Challenge/OpportunityCard.tsx
@@ -5,7 +5,7 @@ import Button from '../core/Button';
 import Card from '../core/Card';
 import { useTranslation } from 'react-i18next';
 import { Nvp } from '../../models/graphql-schema';
-import { Activities } from '../ActivityPanel';
+import { Activities } from '../ActivityPanel/ActivityCard';
 import getActivityCount from '../../utils/get-activity-count';
 import TagContainer from '../core/TagContainer';
 import Tag from '../core/Tag';
